last version

+++ ft_get_image.c
@@ -6,13 +6,13 @@
 /*   By: sbadr <sbadr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/09 15:33:04 by idabligi          #+#    #+#             */
-/*   Updated: 2023/07/25 17:27:26 by sbadr            ###   ########.fr       */
+/*   Updated: 2023/07/26 16:08:06 by sbadr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./includes/cube.h"
 
-uint8_t	get_rgba(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
+int	get_rgba(uint8_t r, uint8_t g, uint8_t b, uint8_t a)
 {
 	return (r << 24 | g << 16 | b << 8 | a);
 }
